Make the `props` of `Position` return a `Try[Props]`
`Position` has the elements needed to validate the passed FEN, as well as the passed initial moves. Whomever calls the `props` method can act upon receiving a `Success`, and can find the new FEN from the passed one and the list of initial moves using some internal method of `Position`.